Entities and Core
We learned how to code business objects and added two interfaces and and a base class. The base class EntityBase is an Abstract class that all the business entities inherit from which also implements the IExtensibleDataObject class allowing for objects to be versionable. The two interface IAccountOwenedEntity helps identify which property has ownership and IIdentifiableEntity that just shows who property is the Key.